New file

+++ 0x15-file_io/0-read_textfile.c
@@ -10,7 +10,7 @@
 
 ssize_t read_textfile(const char *filename, size_t letters)
 {
-	ssize_t fd, num, written_num;
+	int  fd, num, written_num;
 	char *buffer;
 
 	if (!filename)
